+++ server/src/main/java/org/apache/iotdb/db/query/expression/unary/NegationExpression.java
@@ -95,7 +95,7 @@ public class NegationExpression extends Expression {
 
   @Override
   public IntermediateLayer constructIntermediateLayer(
-      int queryId,
+      long queryId,
       UDTFPlan udtfPlan,
       UDFLayer rawTimeSeriesInputLayer,
       Map<Expression, IntermediateLayer> expressionIntermediateLayerMap,
